Make NO Data dynamic  and move storm point tooltip
