+++ react/components/stormPoint.js
@@ -6,7 +6,7 @@ import { remap_coord_array, flip_coords, fetch_value } from "@/lib/storm_utils";
 import {empty_point_obj} from "@/components/storm_point_details"
 import { Tooltip } from 'react-leaflet';
 import { useMediaQuery, useTheme } from '@mui/material';
-import StormPointDetailsTooltip from "./storm_point_details_tooltip";
+import StormPointDetailsTooltip from "./storm_dashboard/storm_point_details_tooltip";
 import { createSvgIconWithText } from "./utils/storm_display_utils";
